#56: refactored ReaderTestSupport renamed ReaderTestSupport to AutocompletionReaderTestSupport
